SBB-1707 remove ranger local docker setup
It is bad practice to store credentials even for local setup in github
repo, especially when the repo is public.

Security team has advised to delete it.

RZP does not rely on this local docker setup, instead has its won setup.
It is simpler to drop this rather than scrub password from each file

Signed-off-by: Utkarsh Saxena <utkarsh.saxena@razorpay.com>